Section_097_-_Day_97_-_Professional_Portfolio_Project_-_[Web_Development]/ecommerce-app/requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-JWT-Extended==4.5.2
python-dotenv==1.0.0
sqlitecloud>=0.0.84
razorpay==1.3.1
Werkzeug==2.3.7
email-validator==2.0.0